[MODIFY] 로그인 기능 구현  - 주석 삭제  - 토큰유효성검사 middleware 분리  - 토큰 만료기간 소스코드 .env 이동  - 에러 핸들링 중복삭제  - 중복 소스코드 및 console 소스코드 삭제
