+++ controllers/userController.js
@@ -28,10 +28,8 @@ const login = async (req, res) => {
       return res.status(400).json({ message: "KEY_ERROR" });
     }
     const accessToken = await userService.login(email, password);
-    if (!accessToken) {
-      return res.status(400).json({ message: "NOT_EXIST_TOKEN" });
-    }
-    return res.status(200).json({ accessToken: accessToken });
+
+    return res.status(200).json({ accessToken });
   } catch (err) {
     console.error(err);
     return res.status(err.statusCode || 500).json({ message: err.message });
