+++ utills/validation-check.js
@@ -1,6 +1,3 @@
-const jwt = require("jsonwebtoken");
-const userDao = require("../models/userDao");
-
 const checkValidationEmail = (email) => {
   const EMAIL_REGEX = new RegExp(
     /^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$/i
@@ -43,34 +40,9 @@ const checkValidationPhoneNumber = (phoneNumber) => {
   }
 };
 
-const checkValidToken = async (req, res, next) => {
-  try {
-    const token = req.headers.authorization;
-    console.log(token);
-    if (!token) {
-      const err = new Error("NOT_EXIST_TOKEN");
-      throw err;
-    }
-    const decoded = jwt.verify(token, process.env.SECRET_KEY);
-    const user = await userDao.checkRegistuserId(decoded.userId);
-    if (!user) {
-      const err = new Error("INVALID_TOKEN");
-      err.statusCode = 400;
-      throw err;
-    }
-    req.user = decoded.userId;
-    next();
-  } catch (err) {
-    console.error(err);
-    err.statusCode = 400;
-    throw err;
-  }
-};
-
 module.exports = {
   checkValidationEmail,
   checkValidationPassword,
   checkValidationBirth,
   checkValidationPhoneNumber,
-  checkValidToken,
 };
